+++ services/merchant/create_test.go
@@ -21,8 +21,8 @@ func TestCreateMerchant(t *testing.T) {
 		category := "SmallRestaurant"
 		imageUrl := "https://bread.com/bread.png"
 		coords := location.Location{
-			Latitude:  helper.ToPointer(float32(6.9)),
-			Longitude: helper.ToPointer(float32(42.0)),
+			Latitude:  helper.ToPointer(6.9),
+			Longitude: helper.ToPointer(42.0),
 		}
 
 		Convey("When called with a valid request", func() {
